switching between 12 and 15 depending on wx3 and wx4 is added
still lighing in 15 is not correct
